test(suite): co-locate utility tests
