+++ packages/suite/src/utils/suite/homescreen.test.ts
@@ -12,7 +12,7 @@ import {
     isValidImageWidth,
 } from 'src/utils/suite/homescreen';
 
-import * as fixtures from '../__fixtures__/homescreen';
+import * as fixtures from '../../../mocks/mockHomescreen';
 
 describe('homescreen', () => {
     describe(fileToDataUrl.name, () => {
