+++ packages/suite/src/utils/wallet/trading/buyUtils.test.ts
@@ -1,7 +1,8 @@
 import { type Account } from '@suite-common/wallet-types';
 
-import { buyUtilsFixtures } from 'src/utils/wallet/trading/__fixtures__/buyUtils';
 import { createQuoteLink, createTxLink, getStatusMessage } from 'src/utils/wallet/trading/buyUtils';
+
+import { buyUtilsFixtures } from '../../../../mocks/mockBuyUtils';
 
 describe('buyUtils', () => {
     const accountMock = {
