+++ packages/suite/src/utils/wallet/trading/sellUtils.test.ts
@@ -1,25 +1,23 @@
-import { type CryptoId, type ExchangeTradeQuoteRequest } from 'invity-api';
-
 import { type TradingComposedTransactionInfo } from '@suite-common/trading';
-import { type Account } from '@suite-common/wallet-types';
 
-import { createQuoteLink } from 'src/utils/wallet/trading/exchangeUtils';
+import { type Account } from 'src/types/wallet';
+import { createQuoteLink } from 'src/utils/wallet/trading/sellUtils';
 
-describe('exchangeUtils', () => {
+import * as fixtures from '../../../../mocks/mockSellUtils';
+
+const { QUOTE_REQUEST_FIAT, QUOTE_REQUEST_CRYPTO } = fixtures;
+
+describe('sellUtils', () => {
     afterEach(() => {
         jest.clearAllMocks();
     });
 
-    const mockQuotesRequest: ExchangeTradeQuoteRequest = {
-        send: 'bitcoin' as CryptoId,
-        receive: 'litecoin' as CryptoId,
-        sendStringAmount: '1',
-    };
     const mockAccount = {
         symbol: 'btc',
         accountType: 'normal',
         index: 1,
     } as Account;
+
     const mockComposedInfo = {
         selectedFee: 'normal',
         composed: {
@@ -29,52 +27,53 @@ describe('exchangeUtils', () => {
             feeLimit: '4',
         },
     } as TradingComposedTransactionInfo;
+
     const mockQuoteId = 'quoteId';
 
     describe('createQuoteLink', () => {
-        it('should create link for quote', async () => {
+        it('should create link for quote for fiat', async () => {
             expect(
                 await createQuoteLink(
-                    mockQuotesRequest,
+                    QUOTE_REQUEST_FIAT,
                     mockAccount,
                     mockComposedInfo,
                     mockQuoteId,
                 ),
             ).toStrictEqual(
-                `${window.location.origin}/coinmarket-redirect#exchange-offers/btc/normal/1/bitcoin/litecoin/1/quoteId/custom/1/2/3/4`,
+                `${window.location.origin}/coinmarket-redirect#sell-offers/btc/normal/1/p-qf/CZ/EUR/10/bitcoin/creditCard/quoteId/custom/1/2/3/4`,
             );
         });
 
         it('should create link for quote when selectedFee is high', async () => {
             expect(
                 await createQuoteLink(
-                    mockQuotesRequest,
+                    QUOTE_REQUEST_CRYPTO,
                     mockAccount,
                     { ...mockComposedInfo, selectedFee: 'high' },
                     mockQuoteId,
                 ),
             ).toStrictEqual(
-                `${window.location.origin}/coinmarket-redirect#exchange-offers/btc/normal/1/bitcoin/litecoin/1/quoteId/custom/1/2/3/4`,
+                `${window.location.origin}/coinmarket-redirect#sell-offers/btc/normal/1/p-qc/CZ/EUR/0.001/bitcoin/creditCard/quoteId/custom/1/2/3/4`,
             );
         });
 
         it('should create link for quote when selectedFee is custom', async () => {
             expect(
                 await createQuoteLink(
-                    mockQuotesRequest,
+                    QUOTE_REQUEST_CRYPTO,
                     mockAccount,
                     { ...mockComposedInfo, selectedFee: 'custom' },
                     mockQuoteId,
                 ),
             ).toStrictEqual(
-                `${window.location.origin}/coinmarket-redirect#exchange-offers/btc/normal/1/bitcoin/litecoin/1/quoteId/custom/1/2/3/4`,
+                `${window.location.origin}/coinmarket-redirect#sell-offers/btc/normal/1/p-qc/CZ/EUR/0.001/bitcoin/creditCard/quoteId/custom/1/2/3/4`,
             );
         });
 
         it('should create link for quote when account network type is solana', async () => {
             expect(
                 await createQuoteLink(
-                    mockQuotesRequest,
+                    QUOTE_REQUEST_CRYPTO,
                     {
                         ...mockAccount,
                         symbol: 'sol',
@@ -84,7 +83,7 @@ describe('exchangeUtils', () => {
                     mockQuoteId,
                 ),
             ).toStrictEqual(
-                `${window.location.origin}/coinmarket-redirect#exchange-offers/sol/normal/1/bitcoin/litecoin/1/quoteId/normal/1/2/3/4`,
+                `${window.location.origin}/coinmarket-redirect#sell-offers/sol/normal/1/p-qc/CZ/EUR/0.001/bitcoin/creditCard/quoteId/normal/1/2/3/4`,
             );
         });
     });
